RAG_LLM/requirements.txt
python-dotenv
langchain[llm]
pypdf
docarray
pandas
giskard[llm]
beautifulsoup4
langchain-openai    
huggingface_hub
sentence-transformers
unstructured[pdf]
openpyxl
huggingface_hub[hf_xet]